update useRepoCoverageTimeseries to TS and to support updated chart
